Adapting to both android and iOS testing.
