+++ lib/screens/event_list.dart
@@ -1,9 +1,11 @@
 import 'dart:convert';
+import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
+import 'package:http/http.dart' as http;
 import 'package:pal_finder/widgets/event_card.dart';
 import 'package:pal_finder/data/event.dart';
-import 'package:pal_finder/core/networking.dart';
+// import 'package:pal_finder/core/networking.dart';
 
 class EventList extends StatefulWidget {
   @override
@@ -17,7 +19,6 @@ class _EventListState extends State<EventList> {
     lng: 151.195697,
     dist: 0.5,
   );
-
 
   // @override
   // void initState() {
@@ -61,5 +62,47 @@ class _EventListState extends State<EventList> {
         },
       ),
     );
+  }
+}
+
+class EventFetcher {
+  EventFetcher({
+    @required this.lat,
+    @required this.lng,
+    @required this.dist,
+  }) {
+    if (Platform.isIOS) {
+      nextUrl='http://localhost:8000/apis/events/nearby/?lng=${lng.toStringAsFixed(6)}&lat=${lat.toStringAsFixed(6)}&dist=${dist.toStringAsFixed(3)}';
+    } else if (Platform.isAndroid) {
+      nextUrl='http://10.0.2.2:8000/apis/events/nearby/?lng=${lng.toStringAsFixed(6)}&lat=${lat.toStringAsFixed(6)}&dist=${dist.toStringAsFixed(3)}';
+    }
+  }
+
+  final double lat;
+  final double lng;
+  final double dist;
+  String nextUrl;
+
+  Future<List<EventData>> fetchNextPage() async {
+    final eventList = <EventData>[];
+    if (nextUrl != null) {
+      final response = await http.get(
+        nextUrl,
+        headers: {
+          HttpHeaders.authorizationHeader: 'Token fc27c2274f456cf0e7bd00663035a8c94ada0025'
+        },
+      );
+      if (response.statusCode == 200) {
+        final data = jsonDecode(response.body) as Map<String, dynamic>;
+        final list = data['results'] as List;
+        list.forEach((eventMap) {
+          eventList.add(EventData.fromMap(eventMap));
+        });
+        nextUrl = data['next'];
+      } else {
+        throw Exception('Fail to load events.');
+      }
+    }
+    return eventList;
   }
 }
